more tests and lint fixes.

+++ x/reward/keeper/reward_claim.go
@@ -18,7 +18,7 @@ func (k Keeper) ClaimRewards(ctx sdk.Context, rewardProgramID uint64, addr strin
 	sent, err := k.SendRewards(ctx, rewardProgram, rewards, addr)
 	if err != nil {
 		// Rollback is handled by the chain automatically
-		//k.RollbackClaims(ctx, rewardProgram, rewards, addr)
+		// k.RollbackClaims(ctx, rewardProgram, rewards, addr)
 		return nil, sdk.Coin{}, err
 	}
 	rewardProgram.ClaimedAmount = rewardProgram.ClaimedAmount.Add(sent)
@@ -42,7 +42,7 @@ func (k Keeper) ClaimRewardsForProgram(ctx sdk.Context, rewardProgram types.Rewa
 }
 
 func (k Keeper) ClaimRewardForPeriod(ctx sdk.Context, rewardProgram types.RewardProgram, period uint64, addr string) (reward types.ClaimedRewardPeriodDetail, found bool) {
-	state, err := k.GetRewardAccountState(ctx, rewardProgram.GetId(), uint64(period), addr)
+	state, err := k.GetRewardAccountState(ctx, rewardProgram.GetId(), period, addr)
 	if err != nil {
 		return reward, false
 	}
@@ -50,14 +50,14 @@ func (k Keeper) ClaimRewardForPeriod(ctx sdk.Context, rewardProgram types.Reward
 		return reward, false
 	}
 
-	distribution, err := k.GetClaimPeriodRewardDistribution(ctx, uint64(period), rewardProgram.GetId())
+	distribution, err := k.GetClaimPeriodRewardDistribution(ctx, period, rewardProgram.GetId())
 	if err != nil {
 		return reward, false
 	}
 
 	participantReward := k.CalculateParticipantReward(ctx, int64(state.GetSharesEarned()), distribution.GetTotalShares(), distribution.GetRewardsPool())
 	reward = types.ClaimedRewardPeriodDetail{
-		ClaimPeriodId:     uint64(period),
+		ClaimPeriodId:     period,
 		TotalShares:       state.GetSharesEarned(),
 		ClaimPeriodReward: participantReward,
 	}
